🚚 rename: 폴더명 직관적으로 변경
